DN-4: Working on test page demonstrating API integration.
